requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6

fastapi==0.128.0
uvicorn==0.40.0

greenlet==3.3.0
Mako==1.3.10
MarkupSafe==3.0.3

psycopg2-binary==2.9.11
SQLAlchemy==2.0.25

pydantic==2.12.5
pydantic_core==2.41.5
typing-inspection==0.4.2
typing_extensions==4.15.0

python-dotenv==1.2.1
PyJWT==2.8.0

# 🔥 Supabase compatibility
supabase==2.3.4
gotrue==2.4.2
postgrest==0.13.2
realtime==1.0.6
storage3==0.7.4
httpx==0.24.1

# Google Calendar integration
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0